breaking: remove drawer and add toolMenu state
